+++ nni/nas/evaluator/__init__.py
@@ -1,0 +1,10 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT license.
+
+from nni.common.framework import shortcut_framework
+
+from .functional import FunctionalEvaluator
+
+shortcut_framework(__name__)
+
+del shortcut_framework
