+++ nni/retiarii/converter/op_types.py
@@ -1,22 +1,4 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from enum import Enum
-
-# except the special case which can not treat as a basic module from pytorch
-MODULE_EXCEPT_LIST = ['Sequential']
-
-
-class OpTypeName(str, Enum):
-    """
-    op type to its type name str
-    """
-    Attr = 'Attr'
-    Constant = 'Constant'
-    LayerChoice = 'LayerChoice'
-    InputChoice = 'InputChoice'
-    ValueChoice = 'ValueChoice'
-    Placeholder = 'Placeholder'
-    MergedSlice = 'MergedSlice'
-    Repeat = 'Repeat'
-    Cell = 'Cell'
+from nni.nas.execution.pytorch.graph.converter.op_types import *
